RL-172: Edit a course

+++ src/components/pages/super-admin/courses/admission-tests/index.tsx
@@ -10,6 +10,7 @@ import { InputText, Text } from 'src/components/shared/ui';
 import CustomTable from 'src/components/shared/ui/table';
 import AdmissionTestTableFilters from 'src/components/shared/ui/table/components/filters/admission-test';
 import { admissionTestHeadCells } from 'src/constants/head-cells';
+import { confirmDelete, invalidForm } from 'src/constants/modal-content';
 import { AdmissionTest } from 'src/interfaces/entities/admission-test';
 import { useAppDispatch, useAppSelector } from 'src/redux';
 import { resetQuery, setQuery } from 'src/redux/modules/admission-test/actions';
@@ -79,53 +80,34 @@ const AdmissionTestsList = () => {
     dispatch(setQuery(`&${new URLSearchParams(data).toString().replace(/_/g, '.')}`));
   };
 
-  const onInputSubmit = async (data) => {
-    if (editId) {
-      const response = await dispatch(
-        editAdmissionTests(editId, {
-          ...data,
-          isActive: true,
-        }),
-      );
-      if (response.error) {
-        dispatch(
-          openModal({
-            title: 'Algo salió mal',
-            description:
-              response.status === 400
-                ? 'El nombre del test de admisión ya esta en uso.'
-                : 'Por favor revise los datos ingresados.',
-            type: 'alert',
+  const onInputSubmit = async (data: { name: string }) => {
+    const response = editId
+      ? await dispatch(
+          editAdmissionTests(editId, {
+            ...data,
+            isActive: true,
+          }),
+        )
+      : await dispatch(
+          createAdmissionTests({
+            ...data,
+            isActive: true,
           }),
         );
-      } else {
-        setValue('name', '');
-        setEditId('');
-        setSelectedObjects([]);
-      }
+    if ('error' in response.payload) {
+      dispatch(
+        openModal({
+          ...invalidForm,
+          description:
+            response.payload.status === 400
+              ? 'El nombre del test de admisión ya esta en uso.'
+              : invalidForm.description,
+        }),
+      );
     } else {
-      const response = await dispatch(
-        createAdmissionTests({
-          ...data,
-          isActive: true,
-        }),
-      );
-      if (response.error) {
-        dispatch(
-          openModal({
-            title: 'Algo salió mal',
-            description:
-              response.status === 400
-                ? 'El nombre del test de admisión ya esta en uso.'
-                : 'Por favor revise los datos ingresados.',
-            type: 'alert',
-          }),
-        );
-      } else {
-        setValue('name', '');
-        setEditId('');
-        setSelectedObjects([]);
-      }
+      setValue('name', '');
+      setEditId('');
+      setSelectedObjects([]);
     }
   };
 
@@ -140,14 +122,14 @@ const AdmissionTestsList = () => {
     setEditId('');
     setSelectedObjects([]);
     dispatch(
-      openModal({
-        title: 'Eliminar test de admisión',
-        description: '¿Está seguro que desea eliminar este test de admisión?',
-        type: 'confirm',
-        handleConfirm: () => {
-          dispatch(deleteAdmissionTest(id));
-        },
-      }),
+      openModal(
+        confirmDelete({
+          entity: 'test de admisión',
+          handleConfirm: () => {
+            dispatch(deleteAdmissionTest(id));
+          },
+        }),
+      ),
     );
   };
 
@@ -162,7 +144,7 @@ const AdmissionTestsList = () => {
     <Box className={styles.container} data-testid="list-admTest-container-div">
       <Box className={styles.toolbarContainer}>
         <Text variant="h1">Tests de admisión</Text>
-        <Text variant="h3" className={styles.subtitle}>
+        <Text variant="subtitle1" className={styles.subtitle}>
           Lista completa con los test de admisión actuales de la aplicacion.
         </Text>
         <div className={styles.toolbar}>
@@ -206,7 +188,7 @@ const AdmissionTestsList = () => {
         </div>
       </Box>
       {errorData.error && errorData.status === 500 ? (
-        <Text data-testid="list-admTest-title-container-div-error" variant="h2">
+        <Text data-testid="list-admTest-title-container-div-error" variant="subtitle1">
           Hubo un error al cargar la tabla de tests de admisión.
         </Text>
       ) : (
